backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
groq==0.11.0
exa-py==1.0.9
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0

